Omstrukturer da activetx bare er for routes
