+++ apps/etterlatte-behandling/src/main/kotlin/SaksbehandlerDao.kt
@@ -1,9 +1,24 @@
 package no.nav.etterlatte
 
 import no.nav.etterlatte.behandling.klienter.SaksbehandlerInfo
+import no.nav.etterlatte.libs.database.toList
 import javax.sql.DataSource
 
 class SaksbehandlerDao(private val dataSource: DataSource) {
+    fun hentalleSaksbehandlere(): List<String> {
+        dataSource.connection.use {
+            val statement =
+                it.prepareStatement(
+                    """
+                    select distinct saksbehandler from oppgave;
+                    """.trimIndent(),
+                )
+            return statement.executeQuery().toList {
+                getString("saksbehandler")
+            }
+        }
+    }
+
     fun upsertSaksbehandler(saksbehandler: SaksbehandlerInfo) {
         dataSource.connection.use {
             val statement =
