+++ apps/etterlatte-behandling/src/main/kotlin/oppgave/OppgaveDao.kt
@@ -18,8 +18,6 @@ import java.sql.ResultSet
 import java.util.UUID
 
 interface OppgaveDao {
-    fun hentAlleIdenter(): List<String>
-
     fun opprettOppgave(oppgaveIntern: OppgaveIntern)
 
     fun hentOppgave(oppgaveId: UUID): OppgaveIntern?
@@ -61,10 +59,6 @@ interface OppgaveDao {
 
 class OppgaveDaoImpl(private val connection: () -> Connection) : OppgaveDao {
     private val logger = LoggerFactory.getLogger(this::class.java)
-
-    override fun hentAlleIdenter(): List<String> {
-        TODO("Not yet implemented") lag en distinct p saksbehandler spørrign
-    }
 
     override fun opprettOppgave(oppgaveIntern: OppgaveIntern) {
         with(connection()) {
